feat: add release date to component page
